refactor: tea composition using suplement container
