+++ src/beverage/coffee.ts
@@ -2,7 +2,6 @@ import { SuplementContainer } from "../suplements/suplement-container";
 import { Beverage } from "./beverage.interface";
 
 export class Coffee implements Beverage {
-
 	constructor(public suplement: SuplementContainer) {}
 
 	price(): number {
