--- src/beverage/tea/tea-with-milk.ts
@@ -1,7 +1,0 @@
-import { Tea } from "./tea";
-
-export class TeaWithMilk extends Tea {
-	public price(): number {
-		return super.price() + 0.1;
-	}
-}
